+++ R/fasano_franceschini_test.R
@@ -6,8 +6,9 @@
 #'
 #' @param S1 a `[n by 2]` `data.frame` of x and y coordinates of sample 1
 #' @param S2 a `[n by 2]` `data.frame` of x and y coordinates of sample 2
-#' @param nBootstrap a `numeric` defining the number of bootstrapped samples to be generated for computing the empirical p-value (note this procedure is slow and computationally expensive on the order of nBootStrap*O(n^2). Default is set to 0.
-#' If nBootstrap is 0, the Fasano Franceschini distributional approximation is used for defining the p-value. See Fasano and Franceschini test (1987) for details.
+#' @param nBootstrap a depreciated `numeric` argument defining the number of bootstrapped samples to be generated for computing the empirical p-value. nBootstrap is set to be replaced by nPermute in the next released version of the package.
+#' @param nPermute a `numeric` defining the number of permuted samples to be generated for computing the empirical p-value (note this procedure is slow and computationally expensive on the order of nPermute*O(n^2). Default is set to 0.
+#' If nPermute is 0, the Fasano Franceschini distributional approximation is used for defining the p-value. See Fasano and Franceschini test (1987) for details.
 #' @param cores a `numeric` defining the number of cores to use for processing
 #'
 #' @return the 2-D ks statistic and p-value
@@ -45,7 +46,7 @@
 #' @import methods
 #' @export
 
-fasano.franceschini.test <- function(S1, S2, nBootstrap = 0, cores = 1) {
+fasano.franceschini.test <- function(S1, S2, nBootstrap = 0, nPermute = 0, cores = 1) {
 
   #validate inputs
   if(!is.data.frame(S1) | !is.data.frame(S2)){
@@ -54,8 +55,16 @@ fasano.franceschini.test <- function(S1, S2, nBootstrap = 0, cores = 1) {
   if(ncol(S1) != 2 | ncol(S2) != 2){
     stop("S1 and S2 must be a `data.frame` of dim nrows x 2 cols, more than 2 cols detected")
   }
-  if(nBootstrap < 0){
-    stop("nBootstrap must be a positive value")
+
+  if (!missing(nBootstrap)) {
+    warning("The 'nBootstrap' argument is deprecated, and will be removed in the next version of this package.
+             Use 'nPermute' instead of 'nBootstrap' to compute the empirical null distribution for your data.
+             The 'nPermute' argument has been set equal to 'nBootstrap'.")
+    nPermute <- nBootstrap
+  }
+
+  if(nPermute < 0){
+    stop("nPermute must be a positive value")
   }
 
   #grab names of samples for output
@@ -77,25 +86,20 @@ fasano.franceschini.test <- function(S1, S2, nBootstrap = 0, cores = 1) {
   names(D) <- "D-stat"
 
 
-  # if bootstrap is enabled, compute the bootstrapped null d
-  if (nBootstrap > 0) {
-    x_marg <- c(S1[, 1], S2[, 1])
-    y_marg <- c(S1[, 2], S2[, 2])
+  # if permutation is enabled, compute the bootstrapped null d
+  if (nPermute > 0) {
+    allData <- rbind(S1,S2)
 
-    d <- parallel::mclapply(X = 1:nBootstrap, mc.cores = cores, FUN = function(i) {
-      S1_resample <- data.frame(
-        x = sample(x = x_marg, size = n1, replace = T),
-        y = sample(x = y_marg, size = n1, replace = T)
-      )
-      S2_resample <- data.frame(
-        x = sample(x = x_marg, size = n2, replace = T),
-        y = sample(x = y_marg, size = n2, replace = T)
-      )
+    d <- parallel::mclapply(X = 1:nPermute, mc.cores = cores, FUN = function(i) {
+      sel_s1 <- sample(1:nrow(allData), size = n1, replace = F)
+      S1_resample <- allData[sel_s1,]
+      S2_resample <- allData[-sel_s1,]
+
       getDstat(originSamples = S1_resample, S1 = S1_resample, S2 = S2_resample, cores = 1)
     })
 
     # count the number of bootstrapped d stats that are larger than the observed
-    pval <- (sum(unlist(d) > D) + 1) / (nBootstrap + 1)
+    pval <- (sum(unlist(d) > D) + 1) / (nPermute + 1)
     names(pval) <- "p-value"
 
   } else {