+++ NEWS.md
@@ -1,3 +1,7 @@
+# fasano.franceschini.test (development version)
+
+* The 'nBootstrap' argument is deprecated, and will be removed in the next version of this package. Use 'nPermute' instead of 'nBootstrap' to compute the empirical null distribution for your data.
+
 # fasano.franceschini.test 1.0.1
 
 * More efficient `quadCount()` calculation reducing runtime.